authorization and authentication implemented using api v1

+++ controllers/api/v1/posts_api.js
@@ -1,5 +1,5 @@
 const Post = require('../../../models/post');
-const Comment=require('../../../models/comment');
+const Comment = require('../../../models/comment');
 
 module.exports.index = async function (req, res)
 {
@@ -35,14 +35,25 @@ module.exports.destroy = async (req, res) =>
     try
     {
         let post = await Post.findById(req.params.id);
+        if (post.user == req.user.id)
+        {
+            post.remove();
+            await Comment.deleteMany({ post: req.params.id });
+            return res.status(200).json(
+                {
+                    message: 'Post and associated comments deleted successfully!'
+                }
+            )
+        }
+        else
+        {
+            return res.status(401).json(
+                {
+                    message:'You cannot delete this post!'
+                }
+            )
+        }
 
-        post.remove();
-        await Comment.deleteMany({ post: req.params.id });
-        return res.status(200).json(
-            {
-                message:'Post and associated comments deleted successfully!'
-            }
-        )
     }
     catch (error)
     {
